Add the Either type
Serves as a way to switch between two virtual nodes of different types.
Could for example be used as an If construct if paired with the empty
virtual node `()`.

With specialization the case where `A, B: SingleNode` could be treated
differently using `Node.replaceChild()`. Therefore, in preparation, the
`node` parameter in `SingleNode::update_node` was made a mutable
reference.